% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeEntourageScore}
\alias{computeEntourageScore}
\title{Computation of the Entourage Score}
\usage{
computeEntourageScore(dist1, dist2, k = 3)
}
\arguments{
\item{dist1}{a numeric matrix of the first distance matrix}

\item{dist2}{a numeric matrix of the second distance matrix}

\item{k}{a numeric indicating the number of nearest neighbours to compare}
}
\value{
a numeric value of the Entourage Score
}
\description{
This function is used to compute the Entourage Score between two distance matrices, which quantifies the local quality of a MDS representation. 

The Entourage Score corresponds to the normalised number of identical nearest neighbours for each object in the two distance matrices.
}
\details{
This function has been implemented in C++ for fast computations and can be executed from R using this wrapper.
}

