% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeKruskalStress}
\alias{computeKruskalStress}
\title{Computation of the Kruskal Stress}
\usage{
computeKruskalStress(dist1, dist2)
}
\arguments{
\item{dist1}{a numeric matrix of the first distance matrix}

\item{dist2}{a numeric matrix of the second distance matrix}
}
\value{
a numeric value the Kruskal Stress
}
\description{
This function is used to compute the Kruskal Stress between two distance matrices, which quantifies the global quality of a MDS representation.
 
The Kruskal Stress corresponds to the quantify of information lost during the dimentionality reduction process.
}
\details{
This function has been implemented in C++ for fast computations and can be executed from R using this wrapper.
}

