% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distManhattan}
\alias{distManhattan}
\title{Computation of a distance matrix using the Manhattan metric}
\usage{
distManhattan(data)
}
\arguments{
\item{data}{a numeric matrix. Rows must correspond to the objects and columns must correspond to the attributes}
}
\value{
a numeric matrix containing the Manhattan distances betweens all objects
}
\description{
This function computes the distance matrix from a numeric matrix using the Manhattan metric. 

Each row of the input matrix must corresponds to an object and each column must corresponds to an attribute.
}
\details{
This function has been implemented in C++ for fast computations and can be executed from R using this wrapper.
}

