% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMDS}
\alias{plotMDS}
\title{Generation of ggplot graphical representations for MDS Reference Maps and MDS Projections.}
\usage{
plotMDS(mds, color = NULL, shape = NULL, polygon = NULL, title = "MDS",
  display.legend = TRUE)
}
\arguments{
\item{mds}{a MDS result provided by the 'MDSReferenceMap()', 'MDSProjection()', or 'MDSReferenceMapwithProjection()' functions}

\item{color}{a vector used to color the points}

\item{shape}{a vector used for the shape of the points}

\item{polygon}{a vector used to color the convex hull}

\item{title}{a character specifying the title of the representation}

\item{display.legend}{a logical specifying if the graphic legend must be displayed}
}
\description{
This function generates a graphical representation for MDS Reference Maps and MDS Projections.
Objects can be shaped and colored using the 'color' and 'shape' parameters.
Convex hulls for given sets of objects can also be displayed.
}

