% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setConstants}
\alias{setConstants}
\title{Setting of the algorithm constants}
\usage{
setConstants(K = 1, F = 0.1, DELTA_T = 0.001, MASS = 10,
  ACC_THRESHOLD = 0)
}
\arguments{
\item{K}{a numeric value indicating the spring strength between two particles}

\item{F}{a numeric value indicating the friction of the springs}

\item{DELTA_T}{a numeric value indicating the time between two steps of the algorithm}

\item{MASS}{a numeric value indicating the mass of each particle}

\item{ACC_THRESHOLD}{a numeric value indicating the acceleration threshold (a value of 0 disactivate this acceleration threshold)}
}
\description{
This function is used to set the different constants of the MDSRefMaps algorithm.
}
\details{
The 'ACC_THRESHOLD' parameter can be used to contrain the object accelerations.
}

